Fix/merge file moved (#17)
* add gene name keys for documentation

* not delete values that already exist

* fix merge problem
